# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VulnLab Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPayload"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthPayload"
...
  /cart:
    get:
      operationId: getCart
      tags: [orders]
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartInput"
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /cart/{productId}:
    delete:
      operationId: removeFromCart
      tags: [orders]
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: checkout
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── CHANNELS & MESSAGES ───────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required: [ok]
    AuthPayload:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        expiresIn:
          type: ["integer", "null"]
      required: [token, user]
    RegisterInput:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required: [username, email, password]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        full_name:
          type: ["string", "null"]
        avatar_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
        balance:
          type: ["number", "null"]
        ssn:
          type: ["string", "null"]
        secret_note:
          type: ["string", "null"]
        password_plain:
          type: ["string", "null"]
        is_active:
          type: boolean
        is_locked:
          type: boolean
        failed_logins:
          type: integer
        last_login:
          type: ["string", "null"]
        created_at:
          type: string
        followers_count:
          type: ["integer", "null"]
        following_count:
          type: ["integer", "null"]
      required: [id, username, role, is_active, is_locked, failed_logins, created_at]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        original_price:
          type: ["number", "null"]
        stock:
          type: integer
        category:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        image_url:
          type: ["string", "null"]
        seller_id:
          type: ["integer", "null"]
        seller:
          $ref: "#/components/schemas/User"
        is_active:
          type: boolean
        rating:
          type: ["number", "null"]
        created_at:
          type: string
      required: [id, name, price, stock, is_active, created_at]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
      required: [id, product_id, quantity]
    CartInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        buyer_id:
          type: integer
        buyer:
          $ref: "#/components/schemas/User"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
        shipping_address:
          type: ["string", "null"]
        created_at:
          type: string
      required: [id, buyer_id, total, status, created_at]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        price_at_purchase:
          type: number
      required: [id, product_id, quantity, price_at_purchase]
    CheckoutInput:
      type: object
      properties:
        shippingAddress:
          type: string
...