# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VulnLab Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { User } from './user';

export interface Order {
  id: number;
  buyer_id: number;
  buyer?: User;
  items?: OrderItem[];
  total: number;
  status: string;
  /** @nullable */
  shipping_address?: string | null;
  created_at: string;
}